# [DC-420] Sr multithread storage systems