--- a/service/src/main/java/bio/terra/catalog/common/RequestContextCopier.java
+++ b/service/src/main/java/bio/terra/catalog/common/RequestContextCopier.java
@@ -1,22 +1,21 @@
 package bio.terra.catalog.common;
 
+import java.util.stream.Stream;
 import org.springframework.web.context.request.RequestAttributes;
 import org.springframework.web.context.request.RequestContextHolder;
 
-import java.util.stream.Stream;
-
 /**
- * This class supports creating a parallel {@code Stream} with a copy of the current thread's
- * Spring request attributes copied to the stream's sub-threads. This has the effect of
- * allowing request scope injected fields to be used in classes that are passed to the sub-threads.
- * <p/>
- * For example, if {@code service} uses a request scope injection, and you want to run it in
+ * This class supports creating a parallel {@code Stream} with a copy of the current thread's Spring
+ * request attributes copied to the stream's sub-threads. This has the effect of allowing request
+ * scope injected fields to be used in classes that are passed to the sub-threads.
+ *
+ * <p>For example, if {@code service} uses a request scope injection, and you want to run it in
  * parallel, you can write:
  *
  * <pre>{@code
- *     int sum = RequestContextCopier.parallelWithRequest(widgets.stream())
- *                      .mapToInt(w -> service.getWeight(w))
- *                      .sum();
+ * int sum = RequestContextCopier.parallelWithRequest(widgets.stream())
+ *                  .mapToInt(w -> service.getWeight(w))
+ *                  .sum();
  * }</pre>
  *
  * @see RequestContextHolder
@@ -27,8 +26,7 @@ public class RequestContextCopier {
 
   private final RequestAttributes attributes = RequestContextHolder.getRequestAttributes();
 
-  private RequestContextCopier() {
-  }
+  private RequestContextCopier() {}
 
   private <T> void copyToThread(T target) {
     if (RequestContextHolder.getRequestAttributes() == null) {
@@ -38,8 +36,8 @@ public class RequestContextCopier {
 
   /**
    * Given a {@code Stream}, make a parallel version of it where the current request scope
-   * attributes are copied to each of the sub-threads. The parallel version is created
-   * using {@link Stream#parallel()}.
+   * attributes are copied to each of the sub-threads. The parallel version is created using {@link
+   * Stream#parallel()}.
    *
    * @param stream the stream to parallelize
    * @return the new parallel stream