--- a/service/src/test/java/bio/terra/catalog/common/RequestContextCopierTest.java
+++ b/service/src/test/java/bio/terra/catalog/common/RequestContextCopierTest.java
@@ -3,8 +3,8 @@ package bio.terra.catalog.common;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.util.function.Function;
 import java.util.stream.Stream;
-import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.config.Scope;
 import org.springframework.mock.web.MockHttpServletRequest;
@@ -13,30 +13,25 @@ import org.springframework.web.context.request.RequestScope;
 import org.springframework.web.context.request.ServletRequestAttributes;
 
 public class RequestContextCopierTest {
+  public static final int THREADS = 2;
 
-  int checkAndSumAttributeValueFromStream(@NotNull Stream<String> stream) {
+  int runTestWithProvider(Function<Stream<String>, Stream<String>> provider) {
+    MockHttpServletRequest request = new MockHttpServletRequest();
+    RequestContextHolder.setRequestAttributes(new ServletRequestAttributes(request));
     Scope scope = new RequestScope();
-    return stream.mapToInt(value -> (int) scope.get("attribute", () -> 1)).sum();
+    return provider
+        .apply(Stream.generate(() -> "ignored").limit(THREADS))
+        .mapToInt(value -> (int) scope.get("attribute", () -> 1))
+        .sum();
   }
 
   @Test
   void setsRequestContextInsideThread() {
-    MockHttpServletRequest request = new MockHttpServletRequest();
-    RequestContextHolder.setRequestAttributes(new ServletRequestAttributes(request));
-    int streamValuesWithAttribute =
-        checkAndSumAttributeValueFromStream(
-            RequestContextCopier.parallelWithRequest(Stream.generate(() -> "ignored").limit(2)));
-    assertEquals(streamValuesWithAttribute, 2);
+    assertEquals(THREADS, runTestWithProvider(RequestContextCopier::parallelWithRequest));
   }
 
   @Test
   void requestContextDoesntFollowToChildThreads() {
-    MockHttpServletRequest request = new MockHttpServletRequest();
-    RequestContextHolder.setRequestAttributes(new ServletRequestAttributes(request));
-    assertThrows(
-        IllegalStateException.class,
-        () ->
-            checkAndSumAttributeValueFromStream(
-                Stream.generate(() -> "ignored").limit(2).parallel()));
+    assertThrows(IllegalStateException.class, () -> runTestWithProvider(Stream::parallel));
   }
 }