Loading the dataset list from the catalog can take a long time. One of the reasons for this is because we were loading the data from underlying storage systems in sequence. This PR moves that to loading the underlying data in parallel, and then unifying once all the data has been fetched. It does this by using stream parallelism. The added complexity in this PR is a result of Spring Request Context not being copied down to child threads. In this case, because we know this request doesn't close until after all the child threads have finished execution, we want to copy down that request context.

Ticket: https://broadworkbench.atlassian.net/browse/DC-420

Timing:
Before:
curl -X 'GET' 'localhost:8080/api/v1/datasets' -H 'accept: application/json'   0.01s user 0.01s system 1% cpu 1.451 total
After:
curl -X 'GET' 'http://localhost:8080/api/v1/datasets' -H  -H   0.01s user 0.01s system 3% cpu 0.445 total