@okotsopoulos 2022-11-30
I will second @petesantos' comment -- the ticket itself has great context but I also find it useful to have it close to the changes, along with a link to the ticket in question.